create observeCanDownloadFiles observer and replace all instances where needed
