+++ app/screens/gallery/document_renderer/index.ts
@@ -3,26 +3,16 @@
 
 import {withDatabase} from '@nozbe/watermelondb/DatabaseProvider';
 import withObservables from '@nozbe/with-observables';
-import {combineLatest, of as of$} from 'rxjs';
-import {switchMap} from 'rxjs/operators';
 
-import {observeConfigBooleanValue, observeLicense} from '@queries/servers/system';
+import {observeCanDownloadFiles} from '@queries/servers/file';
 
 import DocumentRenderer from './document_renderer';
 
 import type {WithDatabaseArgs} from '@typings/database/database';
 
 const enhanced = withObservables([], ({database}: WithDatabaseArgs) => {
-    const enableMobileFileDownload = observeConfigBooleanValue(database, 'EnableMobileFileDownload');
-    const complianceDisabled = observeLicense(database).pipe(
-        switchMap((lcs) => of$(lcs?.IsLicensed === 'false' || lcs?.Compliance === 'false')),
-    );
-    const canDownloadFiles = combineLatest([enableMobileFileDownload, complianceDisabled]).pipe(
-        switchMap(([download, compliance]) => of$(compliance || download)),
-    );
-
     return {
-        canDownloadFiles,
+        canDownloadFiles: observeCanDownloadFiles(database),
     };
 });
